ifcfg: do setup only if needed

+++ nmcli/features/environment.py
@@ -602,12 +602,15 @@ def before_scenario(context, scenario):
                     sleep(2)
                 print ("---------------------------")
             if 'ifcfg-rh' in scenario.tags:
-                print ("---------------------------")
-                print ("setting ifcfg-rh plugin")
-                call("printf '# configured by beaker-test\n[main]\nplugins=ifcfg-rh\n' > /etc/NetworkManager/conf.d/99-xxcustom.conf", shell=True)
-                restart_NM_service()
-                # VV Do not lower this
-                sleep(0.5)
+                if call("NetworkManager --print-config |grep '^plugins=ifcfg-rh'", shell=True) != 0:
+                    print ("---------------------------")
+                    print ("setting ifcfg-rh plugin")
+                    call("printf '# configured by beaker-test\n[main]\nplugins=ifcfg-rh\n' > /etc/NetworkManager/conf.d/99-xxcustom.conf", shell=True)
+                    restart_NM_service()
+                    # VV Do not lower this
+                    sleep(0.5)
+                    if 'simwifi' in scenario.tags:
+                        sleep(5)
             if 'wifi' in scenario.tags:
                 wifi_rescan()
             if 'nmtui_general_activate_screen_no_connections' in scenario.tags:
@@ -969,10 +972,11 @@ def before_scenario(context, scenario):
                 call("nmcli connection delete id testeth0", shell=True)
 
             if 'ifcfg-rh' in scenario.tags:
-                print ("---------------------------")
-                print ("setting ifcfg-rh plugin")
-                call("printf '# configured by beaker-test\n[main]\nplugins=ifcfg-rh\n' > /etc/NetworkManager/conf.d/99-xxcustom.conf", shell=True)
-                restart_NM_service()
+                if call("NetworkManager --print-config |grep '^plugins=ifcfg-rh'", shell=True) != 0:
+                    print ("---------------------------")
+                    print ("setting ifcfg-rh plugin")
+                    call("printf '# configured by beaker-test\n[main]\nplugins=ifcfg-rh\n' > /etc/NetworkManager/conf.d/99-xxcustom.conf", shell=True)
+                    restart_NM_service()
 
             if 'eth3_disconnect' in scenario.tags:
                 print ("---------------------------")
@@ -1634,10 +1638,11 @@ def after_scenario(context, scenario):
                 reset_hwaddr_nmtui('eth2')
                 os.system("sudo ip link del team0")
             if 'ifcfg-rh' in scenario.tags:
-                print ("---------------------------")
-                print ("resetting ifcfg plugin")
-                call('sudo rm -f /etc/NetworkManager/conf.d/99-xxcustom.conf', shell=True)
-                restart_NM_service()
+                if call('test -f /etc/NetworkManager/conf.d/99-xxcustom.conf', shell=True) == 0:
+                    print ("---------------------------")
+                    print ("resetting ifcfg plugin")
+                    call('sudo rm -f /etc/NetworkManager/conf.d/99-xxcustom.conf', shell=True)
+                    restart_NM_service()
             if 'inf' in scenario.tags:
                 os.system("sudo nmcli connection delete id infiniband0 infiniband0-port")
             if 'dsl' in scenario.tags:
@@ -2174,10 +2179,11 @@ def after_scenario(context, scenario):
                 restart_NM_service()
 
             if 'ifcfg-rh' in scenario.tags:
-                print ("---------------------------")
-                print ("resetting ifcfg plugin")
-                call('sudo rm -f /etc/NetworkManager/conf.d/99-xxcustom.conf', shell=True)
-                restart_NM_service()
+                if call('test -f /etc/NetworkManager/conf.d/99-xxcustom.conf', shell=True) == 0:
+                    print ("---------------------------")
+                    print ("resetting ifcfg plugin")
+                    call('sudo rm -f /etc/NetworkManager/conf.d/99-xxcustom.conf', shell=True)
+                    restart_NM_service()
 
             if 'dhcpd' in scenario.tags:
                 print ("---------------------------")
